Remove page.ts and add the functionality to layout.ts
Fix #123
